Refactor ReportSubmitted module
