+++ .eslintrc
@@ -3,7 +3,9 @@
   "extends": ["airbnb"],
   "rules": {
     "react/jsx-filename-extension": [1, { "extensions": [".js", ".jsx"] }],
-    "comma-dangle": ["error", {
+    "comma-dangle": [
+      "error",
+      {
         "functions": "ignore"
       }
     ],
@@ -12,7 +14,7 @@
     "react/prefer-stateless-function": [0, { "ignorePureComponents": true }],
     "max-len": 0,
     "_comment": "remove after https://github.com/babel/babel-eslint/issues/530 fix",
-    "template-curly-spacing" : "off",
+    "template-curly-spacing": "off",
     "indent": "off",
     "consistent-return": 0,
     "react/destructuring-assignment": 0,
@@ -20,7 +22,8 @@
     "react/jsx-one-expression-per-line": 0,
     "operator-linebreak": 0,
     "object-curly-newline": 0,
-    "linebreak-style": 0
+    "linebreak-style": 0,
+    "arrow-parens": [1, "always"]
   },
   "env": {
     "browser": true
